new_requirements.txt
alembic==1.0.11
amqp==2.5.0
backcall==0.1.0
Beaker==1.10.1
beautifulsoup4==4.7.1
billiard==3.6.0.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
cssselect==1.0.3
cssutils==1.0.2
decorator==4.4.0
dogpile.cache==0.7.1
html2text==2018.1.9
hupper==1.8.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
kombu==4.6.3
lxml==4.3.4
Mako==1.0.12
MarkupSafe==1.1.1
mock==3.0.5
mohawk==1.0.0
msgpack-python==0.5.6
nose==1.3.7
oauthlib==3.0.1
parso==0.5.0
PasteDeploy==2.0.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plaster==1.0
plaster-pastedeploy==0.7
premailer==3.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pyramid==1.10.4
pyramid-beaker==0.8
pyramid-debugtoolbar==4.5
pyramid-exclog==1.0
pyramid-handlers==0.5
pyramid-ipython==0.2
pyramid-mailer==0.15.1
pyramid-mako==1.0.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
repoze.lru==0.7
repoze.sendmail==4.4.1
requests==2.22.0
requests-hawk==1.0.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.5
stripe==2.31.0
traitlets==4.3.2
transaction==2.4.0
translationstring==1.3
unstdlib==1.7
urllib3==1.25.3
venusian==1.2.0
vine==1.3.0
waitress==1.3.0
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
zope.deprecation==4.4.0
zope.interface==4.6.0
